+++ NEWS.md
@@ -1,6 +1,8 @@
 
 ### acefile 0.5.3-dev
 
+-   Remove the AceFile alias of the AceArchive class.
+-   Rename UnknownMethodError to UnknownCompressionMethodError.
 -   Change AceArchive.mtime and AceMember.mtime to .datetime in order to avoid
     confusion as ACE does not have separate modification and creation times.
 -   Change AceMember.orig_filename (str) to AceMember.raw_filename (bytes).